Creating SimplifiedLogin DataModels
