+++ Sources/AccountSDKIOSWeb/Simplified Login/ViewModel/SimplifiedLoginViewModel.swift
@@ -1,36 +1,6 @@
 import Foundation
 import AuthenticationServices
 
-protocol SimplifiedLoginViewModelUserData {
-    var userContext: UserContextFromTokenResponse { get }
-    var userProfileResponse: UserProfileResponse { get }
-    var displayName: String { get }
-    var initials: String { get }
-}
-
-protocol SimplifiedLoginNamedImageData {
-    var env: ClientConfiguration.Environment { get }
-    var iconNames: [String] { get }
-    var schibstedLogoName: String { get }
-}
-
-extension SimplifiedLoginNamedImageData {
-    var iconNames: [String] {
-        let orderedIconNames: [String]
-        switch env {
-        case .proCom:
-            orderedIconNames = ["Blocket", "Aftonbladet", "SVD", "Omni", "TvNu"]
-        case .proNo:
-            orderedIconNames = ["Finn", "VG", "Aftenposten", "E24", "BergensTidene"]
-        case .proFi:
-            orderedIconNames = ["Tori", "Oikotie", "Hintaopas", "Lendo", "Rakentaja"]
-        case .proDk, .pre:
-            orderedIconNames = ["Blocket", "Aftonbladet", "SVD", "Omni", "TvNu"] // Swedish icons as default
-        }
-        return orderedIconNames
-    }
-}
-
 protocol SimplifiedLoginViewModelAuthenticator {
     var asWebAuthenticationSession: ASWebAuthenticationSession? { get set }
 }
@@ -58,38 +28,31 @@ extension SimplifiedLoginUserActionable {
     }
 }
 
-class SimplifiedLoginViewModel: SimplifiedLoginUserActionable, SimplifiedLoginViewModelUserData, SimplifiedLoginNamedImageData, SimplifiedLoginViewModelAuthenticator {
+class SimplifiedLoginViewModel: SimplifiedLoginUserActionable, SimplifiedLoginViewModelAuthenticator {
     
     let clientName = "Finn" // TODO: Need to be fetched
     
-    init(env: ClientConfiguration.Environment, userContext: UserContextFromTokenResponse, userProfileResponse: UserProfileResponse) {
-        self.env = env
-        self.userContext = userContext
-        self.userProfileResponse = userProfileResponse
+    let imageDataModel: SimplifiedLoginNamedImageData
+    var schibstedLogoName: String { return imageDataModel.schibstedLogoName }
+    var iconNames: [String] { return imageDataModel.iconNames }
+    
+    let userData: SimplifiedLoginViewModelUserData
+    var displayName: String { return userData.userContext.display_text }
+    var initials: String {
+        let firstName  = userData.userProfileResponse.name?.givenName ?? ""
+        let lastName = userData.userProfileResponse.name?.familyName ?? ""
+        let initials = "\(firstName.first?.uppercased() ?? "")\(lastName.first?.uppercased() ?? "")"
+        return initials
+    }
+    
+    init(imageDataModel: SimplifiedLoginNamedImageData, userDataModel: SimplifiedLoginViewModelUserData) {
+        self.imageDataModel = imageDataModel
+        self.userData = userDataModel
     }
     
     // MARK: SimplifiedLoginViewModelAuthenticator
     
     var asWebAuthenticationSession: ASWebAuthenticationSession?
-    
-    // MARK: SimplifiedLoginNamedImageData
-    
-    var env: ClientConfiguration.Environment
-    var schibstedLogoName: String = "sch-logo"
-    
-    // MARK: SimplifiedLoginViewModelUserData
-    
-    let userContext: UserContextFromTokenResponse
-    let userProfileResponse: UserProfileResponse
-    var displayName: String {
-        return userContext.display_text
-    }
-    var initials: String {
-        let firstName  = userProfileResponse.name?.givenName ?? ""
-        let lastName = userProfileResponse.name?.familyName ?? ""
-        let initials = "\(firstName.first?.uppercased() ?? "")\(lastName.first?.uppercased() ?? "")"
-        return initials
-    }
 
     // MARK: SimplifiedLoginUserActionableSimplified Login User Actions
